refactor(slack): split modules
