+++ apps/backend/src/automation/actions/sendSlackMessage/index.e2e.test.ts
@@ -1,11 +1,11 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { factory, setupDatabase } from "@/database/testing";
-import { postMessageToSlackChannel } from "@/slack";
+import { postMessageToSlackChannel } from "@/slack/channel.js";
 
 import { automationAction } from "./index";
 
-vi.mock("@/slack", () => ({
+vi.mock("@/slack/channel.js", () => ({
   __esModule: true,
   postMessageToSlackChannel: vi.fn(),
 }));
